some geometry not rendering

+++ utility_funcs.cpp
@@ -574,9 +574,9 @@ vector< vector<vec3> > getTriangles(const vector<float> &vec_vertices, int offse
 	return triangles;
 }
 
-vector< vector<vec3> > getTriangles(const vector<float> &vec_vertices, const vector<unsigned int> &vertex_indices, int offset, int stride)
+vector< vector<vec3> > getTriangles(const vector<float> &vec_vertices, const vector<unsigned short> &vertex_indices, int offset, int stride)
 {
-	map<unsigned int, vector<float> > vertex_map;
+	map<unsigned short, vector<float> > vertex_map;
 	vector<vec3> vertices;
 	//stride output from obj parser is in bit size, not count
 	int stride_count = stride / sizeof(float);
